Rebrand library to LethalAPI
