--- LCAPI.TerminalCommands.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34031.279
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LCAPI.TerminalCommands", "LCAPI.TerminalCommands\LCAPI.TerminalCommands.csproj", "{C3B419FF-E0B8-486A-96CC-0DB1B72FFD10}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C3B419FF-E0B8-486A-96CC-0DB1B72FFD10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C3B419FF-E0B8-486A-96CC-0DB1B72FFD10}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C3B419FF-E0B8-486A-96CC-0DB1B72FFD10}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C3B419FF-E0B8-486A-96CC-0DB1B72FFD10}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D3ED42EB-28A9-414D-8BED-74E9D2389AE7}
-	EndGlobalSection
-EndGlobal
