--- LCAPI.TerminalCommands/Attributes/RemainingTextAttribute.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace LCAPI.TerminalCommands.Attributes
-{
-	/// <summary>
-	/// Designates that a string argument should contain the remaining text in the command
-	/// </summary>
-	[AttributeUsage(AttributeTargets.Parameter)]
-	public sealed class RemainingTextAttribute : Attribute
-	{
-	}
-}
