--- LCAPI.TerminalCommands/Models/PersistType.cs
@@ -1,9 +1,0 @@
-﻿namespace LCAPI.TerminalCommands.Models
-{
-	public enum PersistType
-	{
-		LocalPlayer,
-		Host,
-		Save
-	}
-}
