+++ LethalAPI.TerminalCommands.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34031.279
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LethalAPI.TerminalCommands", "LethalAPI.TerminalCommands\LethalAPI.TerminalCommands.csproj", "{C6140051-73BA-4EA4-B012-F85F24707185}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C6140051-73BA-4EA4-B012-F85F24707185}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6140051-73BA-4EA4-B012-F85F24707185}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6140051-73BA-4EA4-B012-F85F24707185}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6140051-73BA-4EA4-B012-F85F24707185}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D3ED42EB-28A9-414D-8BED-74E9D2389AE7}
+	EndGlobalSection
+EndGlobal
