+++ LethalAPI.TerminalCommands/Models/PersistType.cs
@@ -1,0 +1,9 @@
+﻿namespace LethalAPI.TerminalCommands.Models
+{
+	public enum PersistType
+	{
+		LocalPlayer,
+		Host,
+		Save
+	}
+}
