+++ LethalAPI.TerminalCommands/NuGet.Config
